McpFC/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-assistant"
version = "0.1.0"
description = "基于MCP和OpenAI API的代码助手工具"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Code Assistant Team"}
]
dependencies = [
    "openai>=1.0.0",
    "mcp>=1.2.0",
    "modelcontextprotocol>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ["py37"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"]
